sialyh/requirements.txt
asgiref==3.7.2	
crispy-bootstrap4==2022.1
Django==3.2
django-crispy-forms==2.0
django-filter==23.5
django-tables2==2.7.0
Pillow==9.5.0
psycopg2==2.9.9
pytz==2024.1	
sqlparse==0.4.4	
typing-extensions==4.7.1
PyMySQL==1.1.0
openpyxl==3.1.2
pandas
